Move scrollX/scrollY to browser-rum=core
