+++ packages/rum-core/src/index.ts
@@ -26,6 +26,7 @@ export { ViewContexts, ViewContext } from './domain/contexts/viewContexts'
 export { RumSessionManager, RumSessionPlan, RumSession } from './domain/rumSessionManager'
 export { getMutationObserverConstructor } from './browser/domMutationObservable'
 export { initViewportObservable, getViewportDimension } from './browser/viewportObservable'
+export { getScrollX, getScrollY } from './browser/scroll'
 export { RumInitConfiguration, RumConfiguration } from './domain/configuration'
 export { DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE } from './domain/rumEventsCollection/action/getActionNameFromElement'
 export { STABLE_ATTRIBUTES } from './domain/rumEventsCollection/action/getSelectorFromElement'
