+++ packages/rum/src/domain/record/observers/scrollObserver.ts
@@ -1,10 +1,10 @@
 import type { DefaultPrivacyLevel, ListenerHandler } from '@datadog/browser-core'
 import { DOM_EVENT, throttle, addEventListener } from '@datadog/browser-core'
+import { getScrollX, getScrollY } from '@datadog/browser-rum-core'
 import type { ElementsScrollPositions } from '../elementsScrollPositions'
 import { getEventTarget } from '../eventsUtils'
 import { getNodePrivacyLevel } from '../privacy'
 import { getSerializedNodeId, hasSerializedNode } from '../serialization'
-import { getScrollX, getScrollY } from '../viewports'
 import type { ScrollPosition } from '../../../types'
 import { NodePrivacyLevel } from '../../../constants'
